+++ app/controllers/periodictask_controller.rb
@@ -245,7 +245,7 @@ class PeriodictaskController < ApplicationController
       :due_date_number, :due_date_units, :description, :issue_category_id,
       :estimated_hours, :checklists_template_id, :parent_id, :priority_id, :status_id, :done_ratio, :tag_list,
       :fixed_version_id, :is_active,
-      :monthly_mode,
+      :monthly_mode, :weekend_adjustment,
       { weekdays: [] },
       { month_weeks: [] },
       { custom_field_values: {} },